backend/requirements.txt
attrs==25.3.0
azureopenai==0.0.1
blinker==1.9.0
click==8.1.8
cryptography==45.0.7
descope==1.7.9
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
email-validator==2.3.0
exceptiongroup==1.2.2
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.59.2
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.70.0
httpx==0.28.1
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.4
langsmith==0.4.21
liccheck==0.9.2
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
multidict==6.6.4
networkx==3.5
numpy==2.3.2
openai==1.102.0
orjson==3.11.3
packaging==25.0
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
reportlab==4.4.3
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0

